+++ collections/arrayList/arrayList.go
