+++ collections/forwardList/forwardList.go
