+++ collections/forwardList/forwardList_test.go
