+++ collections/hashSet/hashSet.go
