+++ collections/linkedList/linkedList.go
