+++ collections/maybe/maybe.go
