--- src/interfaces/foldable.go
@@ -1,4 +1,0 @@
-package interfaces
-
-type Foldable[T any] interface {
-}
