+++ stream/all.go
