+++ stream/any.go
