+++ stream/drop.go
