+++ stream/dropWhile.go
