+++ stream/drop_test.go
