+++ stream/error.go
