+++ stream/filter.go
