+++ stream/flatMap.go
