+++ stream/flatMap_test.go
