+++ stream/fmap.go
