+++ stream/fmap_test.go
