+++ stream/reduce.go
