+++ stream/reduce_test.go
