+++ stream/take.go
