+++ stream/takeWhile.go
