+++ stream/takeWhile_test.go
