+++ stream/take_test.go
